refactor(nestjs-signed-url): move gateway contract to interfaces
